add etherchain to transaction succeeded notification
